refactor(FE): App 컴포넌트에서 라우팅 및 인증 흐름 리팩토링
- 라우터 설정을 main.tsx으로 이동하여 구조 개선
- Redux를 활용하여 LoginRequiredRoute의 인증 상태 관리 업데이트
- Header 컴포넌트에서 isAuth prop을 제거하고 userInfo를 직접 사용하도록 단순화
- AuthError 및 ConnectionError 컴포넌트를 통해 라우트의 에러 처리 및 사용자 경험 개선
