+++ front/src/App.tsx
@@ -5,35 +5,11 @@ import { setScrollState } from "@/state/store/scrollStateSlice";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { UserInfo } from "@/shared/types/interface";
 import { getLoginUserInfo } from "@/api/auth";
-import useAuth from "@/shared/hooks/useAuth";
-import Header from "@/widgets/Header";
-import ReviewDetailModal from "@/pages/ReviewDetailModal";
-import SearchDialog from "@/features/common/SearchDialog";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import LoginRequiredRoute from "./pages/LoginRequiredRoute";
-import Feed from "@/pages/Feed";
-import Test from "@/pages/Test";
-import WriteReview from "@/pages/WriteReview";
-import Authorization from "@/pages/Authorization";
-import Onboarding from "@/pages/Onboarding";
-import EditReview from "@/pages/EditReview";
-import Profile from "@/pages/Profile";
-import Notification from "@/pages/Notification";
-import Search from "@/pages/Search";
 import { API_URL } from "@/shared/constants";
-import { fetchNotifications } from "./api/notification";
-import { fetchLatestFeed, fetchPopularFeed } from "./api/review";
-import NotFoundError from "./pages/NotFoundError";
-import ConnectionError from "./pages/ConnectionError";
-import PostsTab from "@/widgets/PostsTab";
-import CommentsTab from "@/widgets/CommentsTab";
-import LikedTab from "@/widgets/LikedTab";
-import {
-  fetchUserCommentList,
-  fetchUserInfoById,
-  fetchUserLikedList,
-  fetchUserReviewList,
-} from "./api/user";
+import useAuth from "@/shared/hooks/useAuth";
+import Header from "./widgets/Header";
+import ReviewDetailModal from "./pages/ReviewDetailModal";
+import SearchDialog from "./features/common/SearchDialog";
 
 function App() {
   // 새로고침 시 로그인 유지를 위해 사용자 정보 조회
@@ -50,161 +26,6 @@ function App() {
     refetchOnWindowFocus: false,
     enabled: isAuth,
   });
-
-  const router = createBrowserRouter([
-    {
-      element: (
-        <>
-          <Header isAuth={isAuth} />
-          <ReviewDetailModal />
-          <SearchDialog />
-        </>
-      ),
-      children: [
-        {
-          path: "/",
-          element: <Feed />,
-          errorElement: <ConnectionError />,
-          loader: async () => {
-            const initialData = await fetchLatestFeed("");
-            return initialData;
-          },
-        },
-        {
-          path: "/latest",
-          element: <Feed />,
-          errorElement: <ConnectionError />,
-          loader: async () => {
-            const initialData = await fetchLatestFeed("");
-            return initialData;
-          },
-        },
-        {
-          path: "/popular",
-          element: <Feed />,
-          errorElement: <ConnectionError />,
-          loader: async () => {
-            const initialData = await fetchPopularFeed("");
-            return initialData;
-          },
-        },
-        {
-          path: "/write",
-          element: (
-            <LoginRequiredRoute isAuth={isAuth}>
-              <WriteReview />
-            </LoginRequiredRoute>
-          ),
-        },
-        {
-          path: "/edit",
-          element: (
-            <LoginRequiredRoute isAuth={isAuth}>
-              <EditReview />
-            </LoginRequiredRoute>
-          ),
-        },
-        {
-          path: "/profile/:id",
-          element: <Profile />,
-          loader: async ({ params }) => {
-            const initialData = await fetchUserInfoById(Number(params.id));
-            return initialData;
-          },
-          errorElement: <NotFoundError />,
-          children: [
-            {
-              path: "",
-              loader: async ({ params }) => {
-                const initialData = await fetchUserReviewList(
-                  Number(params.id),
-                  ""
-                );
-                return initialData;
-              },
-              element: <PostsTab />,
-            },
-            {
-              path: "posts",
-              loader: async ({ params }) => {
-                const initialData = await fetchUserReviewList(
-                  Number(params.id),
-                  ""
-                );
-                return initialData;
-              },
-              element: <PostsTab />,
-            },
-            {
-              path: "comments",
-              loader: async ({ params }) => {
-                const initialData = await fetchUserCommentList(
-                  Number(params.id)
-                );
-                return initialData;
-              },
-              element: <CommentsTab />,
-            },
-            {
-              path: "liked",
-              loader: async ({ params }) => {
-                const initialData = await fetchUserLikedList(
-                  Number(params.id),
-                  ""
-                );
-                return initialData;
-              },
-              element: <LikedTab />,
-            },
-          ],
-        },
-        {
-          path: "/notifications",
-          loader: async () => {
-            try {
-              if (isAuth) {
-                const notifications = await fetchNotifications();
-                return notifications;
-              }
-
-              return [];
-            } catch (error) {
-              return [];
-            }
-          },
-          element: (
-            <LoginRequiredRoute isAuth={isAuth}>
-              <Notification />
-            </LoginRequiredRoute>
-          ),
-        },
-        {
-          path: "/search",
-          element: <Search />,
-        },
-        {
-          path: "/test",
-          element: <Test />,
-        },
-        {
-          path: "/oauth/kakao",
-          element: <Authorization />,
-        },
-        {
-          path: "/onboarding",
-          element: (
-            <LoginRequiredRoute isAuth={isAuth}>
-              <Onboarding />
-            </LoginRequiredRoute>
-          ),
-        },
-        {
-          path: "*",
-          element: <NotFoundError />,
-        },
-      ],
-    },
-  ]);
 
   useEffect(() => {
     if (isAuth) {
@@ -251,7 +72,13 @@ function App() {
     }
   }, [isAuth, queryClient]);
 
-  return <RouterProvider router={router} />;
+  return (
+    <>
+      <Header />
+      <ReviewDetailModal />
+      <SearchDialog />
+    </>
+  );
 }
 
 export default App;
