+++ front/src/widgets/Header.tsx
@@ -11,7 +11,7 @@ import SearchBar from "@/features/common/SearchBar";
 import TooltipWrapper from "@/shared/original-ui/TooltipWrapper";
 import WriteReviewButton from "@/features/review/WriteReviewButton";
 
-export default function Header({ isAuth }: { isAuth: boolean }) {
+export default function Header() {
   const userInfo = useSelector((state: RootState) => state.userInfo);
   const isScrollingUp = useSelector(
     (state: RootState) => state.scrollState.isScrollingUp
@@ -33,7 +33,7 @@ export default function Header({ isAuth }: { isAuth: boolean }) {
             <div className="flex flex-1 md:justify-end items-center gap-2">
               <SearchBar />
               <ThemeToggleButton />
-              {isAuth ? (
+              {userInfo.kakaoId ? (
                 // 로그인을 했을 경우 유저가 사용 가능한 버튼 보여주기
                 <>
                   <WriteReviewButton />
